YouTube Errors refactor
